// src/app/core/types/reports/BundledUserWiseActivityTracker.ts
export interface BundledUserWiseActivityTracker {
    assignee?: string;
    userName?: string;
    docCheck?: number;
    aipDocCheck?: number;
    ebDataEntry?: number;
    frmAlertChecker?: number;
    relationshipMaker?: number;
    relationshipChecker?: number;
    ccplEntryMaker?: number;
    ccplEntryChecker?: number;
    underwriter?: number;
    cboAlChecker?: number;
    cboAlMaker?: number;
    cboPlChecker?: number;
    cboPlMaker?: number;
    topUpChecker?: number;
    topUpMaker?: number;
    pendUser?: number;
    cellUser?: number;
}